Rename transitions.js
